backend/fix: changed cancellaiton dues to double in tables

+++ Backend/dev/migrations/dynamic-offer-driver-app/0320-cancellationDues-column-in-fareparams.sql
@@ -20,6 +20,6 @@ CREATE TABLE atlas_driver_offer_bpp.cancellation_charges (
 
 CREATE INDEX ON atlas_driver_offer_bpp.cancellation_charges USING btree (driver_id);
 
-ALTER TABLE atlas_driver_offer_bpp.search_request_for_driver ADD column customer_cancellation_dues int DEFAULT 0 NOT NULL;
+ALTER TABLE atlas_driver_offer_bpp.search_request_for_driver ADD column customer_cancellation_dues double precision DEFAULT 0 NOT NULL;
 
-ALTER TABLE atlas_driver_offer_bpp.search_request ADD column customer_cancellation_dues int DEFAULT 0 NOT NULL;
+ALTER TABLE atlas_driver_offer_bpp.search_request ADD column customer_cancellation_dues double precision DEFAULT 0 NOT NULL;
